Beamify TestReadMessages
